fix ut and adjust tree-delta-map encoder
