+++ libs/cosmos-sdk/store/rootmulti/store_encoder_test.go
@@ -9,7 +9,7 @@ import (
 	dbm "github.com/tendermint/tm-db"
 )
 
-func newTestTreeDelta() map[string]iavltree.TreeDelta {
+func newTestTreeDeltaMap() iavltree.TreeDeltaMap {
 	// new multiplex store
 	var db dbm.DB = dbm.NewMemDB()
 	ms := newMultiStoreWithMounts(db, types.PruneNothing)
@@ -26,14 +26,8 @@ func newTestTreeDelta() map[string]iavltree.TreeDelta {
 	store2.Set(k2, v2)
 
 	// each store to be committed and return its delta
-	returnedDeltas := make(map[string]iavltree.TreeDelta)
-	for key, store := range ms.stores {
-		_, reDelta, _ := store.Commit(nil, nil)
-		if store.GetStoreType() == types.StoreTypeTransient {
-			continue
-		}
-		returnedDeltas[key.Name()] = reDelta
-	}
+	_, returnedDeltas := ms.CommitterCommitMap(nil)
+
 	return returnedDeltas
 }
 
@@ -41,7 +35,7 @@ func newTestTreeDelta() map[string]iavltree.TreeDelta {
 func TestAminoEncodeDelta(t *testing.T) { testEncodeTreeDelta(t, newEncoder("amino")) }
 func TestJsonEncodeDelta(t *testing.T)  { testEncodeTreeDelta(t, newEncoder("json")) }
 func testEncodeTreeDelta(t *testing.T, enc encoder) {
-	deltaList := newTestTreeDelta()
+	deltaList := newTestTreeDeltaMap()
 
 	_, err := enc.encodeFunc(deltaList)
 	require.NoError(t, err, enc.name())
@@ -51,7 +45,7 @@ func testEncodeTreeDelta(t *testing.T, enc encoder) {
 func BenchmarkAminoEncodeDelta(b *testing.B) { benchmarkEncodeDelta(b, newEncoder("amino")) }
 func BenchmarkJsonEncodeDelta(b *testing.B)  { benchmarkEncodeDelta(b, newEncoder("json")) }
 func benchmarkEncodeDelta(b *testing.B, enc encoder) {
-	data := newTestTreeDelta()
+	data := newTestTreeDeltaMap()
 
 	b.ReportAllocs()
 	b.ResetTimer()
@@ -65,7 +59,7 @@ func benchmarkEncodeDelta(b *testing.B, enc encoder) {
 func TestAminoDecodeDelta(t *testing.T) { testDecodeTreeDelta(t, newEncoder("amino")) }
 func TestJsonDecodeDelta(t *testing.T)  { testDecodeTreeDelta(t, newEncoder("json")) }
 func testDecodeTreeDelta(t *testing.T, enc encoder) {
-	deltaList1 := newTestTreeDelta()
+	deltaList1 := newTestTreeDeltaMap()
 	data, err := enc.encodeFunc(deltaList1)
 
 	_, err = enc.decodeFunc(data)
@@ -76,7 +70,7 @@ func testDecodeTreeDelta(t *testing.T, enc encoder) {
 func BenchmarkAminoDecodeDelta(b *testing.B) { benchmarkDecodeDelta(b, newEncoder("amino")) }
 func BenchmarkJsonDecodeDelta(b *testing.B)  { benchmarkDecodeDelta(b, newEncoder("json")) }
 func benchmarkDecodeDelta(b *testing.B, enc encoder) {
-	deltaList1 := newTestTreeDelta()
+	deltaList1 := newTestTreeDeltaMap()
 	data, _ := enc.encodeFunc(deltaList1)
 
 	b.ReportAllocs()
@@ -88,8 +82,8 @@ func benchmarkDecodeDelta(b *testing.B, enc encoder) {
 
 type encoder interface {
 	name() string
-	encodeFunc(map[string]iavltree.TreeDelta) ([]byte, error)
-	decodeFunc([]byte) (map[string]*iavltree.TreeDelta, error)
+	encodeFunc(iavltree.TreeDeltaMap) ([]byte, error)
+	decodeFunc([]byte) (iavltree.TreeDeltaMap, error)
 }
 
 func newEncoder(encType string) encoder {
@@ -107,22 +101,22 @@ func newEncoder(encType string) encoder {
 type aminoEncoder struct{}
 
 func (ae *aminoEncoder) name() string { return "amino" }
-func (ae *aminoEncoder) encodeFunc(data map[string]iavltree.TreeDelta) ([]byte, error) {
-	return MarshalAppliedDeltaToAmino(data)
+func (ae *aminoEncoder) encodeFunc(data iavltree.TreeDeltaMap) ([]byte, error) {
+	return iavltree.MarshalTreeDeltaMapToAmino(data)
 }
-func (ae *aminoEncoder) decodeFunc(data []byte) (map[string]*iavltree.TreeDelta, error) {
-	return UnmarshalAppliedDeltaFromAmino(data)
+func (ae *aminoEncoder) decodeFunc(data []byte) (iavltree.TreeDeltaMap, error) {
+	return iavltree.UnmarshalTreeDeltaMapFromAmino(data)
 }
 
 // json encoder
 type jsonEncoder struct{}
 
 func (je *jsonEncoder) name() string { return "json" }
-func (je *jsonEncoder) encodeFunc(data map[string]iavltree.TreeDelta) ([]byte, error) {
+func (je *jsonEncoder) encodeFunc(data iavltree.TreeDeltaMap) ([]byte, error) {
 	return itjs.Marshal(data)
 }
-func (je *jsonEncoder) decodeFunc(data []byte) (map[string]*iavltree.TreeDelta, error) {
-	deltalist := map[string]*iavltree.TreeDelta{}
-	err := itjs.Unmarshal(data, &deltalist)
-	return deltalist, err
+func (je *jsonEncoder) decodeFunc(data []byte) (iavltree.TreeDeltaMap, error) {
+	deltaList := map[string]*iavltree.TreeDelta{}
+	err := itjs.Unmarshal(data, &deltaList)
+	return deltaList, err
 }
